check-graphite: can now check most recent point
Instead of just a percent, it's possible to say that you only care about
the most recent point. This is useful because of the way that graphite
APIs expose data, or rather: don't.

Data isn't exposed until the caches flush it to disk which can take an
arbitrary amount of time. So it makes sense to have an arbitrarily large
window of data to examine. But if you are checking the last 30 minutes
worth of data that means that it can take up to 30 minutes for a
transitory incident to resolve.

If you are monitoring something that you don't care about getting
alerted for if it resolves itself, it makes sense to use the `most
recent` assertion form. If you want to get alerted no matter what, you
should not use it.