check-cpu: add active+iowait and active-nice types
These aggregate types make it easier to check for common things you
might care about.

Active + IOWait is useful if you want an alert that is more similar to
check-load, but with more details about exactly what's happening.

Active - Nice is useful if you've got some background process that is
niced that takes up a lot of CPU but you know about it and don't care.