check-ram: fix per-proc percents and show usage
Now the check shows the real amount of RAM used by the process, as well
as correctly showing the per-process percent. Before I was calculating
the ratio of Pages to KiB, which is not correct.

Obviously we need type-validation of the units of those too, as well as
actually asserting that the values read from /proc/meminfo are KiB,
since technically they can change.