[check-procs] Simplify the filter_procs function
I assume I was trying to optimize the number of comparisons executed by only
adding filters to the chain if they were necessary, but this implementation is
so much more reasonable.

And since I didn't benchmark, there's no real reason to believe that the
pointer chasing wouldn't have been a pessimization compared to "this just
always check if we need to apply the filter" style.